refactor: render dependency template with runtime template
