+++ crates/rspack_plugin_javascript/src/dependency/amd/amd_require_item_dependency.rs
@@ -2,7 +2,7 @@ use rspack_cacheable::{cacheable, cacheable_dyn, with::AsPreset};
 use rspack_core::{
   AffectType, AsContextDependency, Dependency, DependencyCategory, DependencyCodeGeneration,
   DependencyId, DependencyRange, DependencyTemplate, DependencyTemplateType, DependencyType,
-  FactorizeInfo, ModuleDependency, TemplateContext, TemplateReplaceSource, module_raw,
+  FactorizeInfo, ModuleDependency, TemplateContext, TemplateReplaceSource,
 };
 use rspack_util::atom::Atom;
 
@@ -110,13 +110,16 @@ impl DependencyTemplate for AMDRequireItemDependencyTemplate {
       return;
     };
     // ModuleDependencyTemplateAsRequireId
-    let content = module_raw(
-      code_generatable_context.compilation,
-      code_generatable_context.runtime_requirements,
-      &dep.id,
-      &dep.request,
-      dep.weak(),
-    );
+    let content = code_generatable_context
+      .compilation
+      .runtime_template
+      .module_raw(
+        code_generatable_context.compilation,
+        code_generatable_context.runtime_requirements,
+        &dep.id,
+        &dep.request,
+        dep.weak(),
+      );
     source.replace(range.start, range.end, &content, None);
   }
 }
