+++ crates/rspack_plugin_javascript/src/dependency/commonjs/common_js_export_require_dependency.rs
@@ -11,7 +11,7 @@ use rspack_core::{
   ModuleGraphCacheArtifact, ModuleIdentifier, Nullable, PrefetchExportsInfoMode, ReferencedExport,
   RuntimeGlobals, RuntimeSpec, TemplateContext, TemplateReplaceSource, UsageState, UsedName,
   collect_referenced_export_items, create_exports_object_referenced, create_no_exports_referenced,
-  module_raw, property_access, to_normal_comment,
+  property_access, to_normal_comment,
 };
 use rustc_hash::FxHashSet;
 use swc_core::atoms::Atom;
@@ -503,7 +503,7 @@ impl DependencyTemplate for CommonJsExportRequireDependencyTemplate {
         UsedName::Normal(used_imported) => {
           format!(
             "{}{}{}",
-            module_raw(
+            compilation.runtime_template.module_raw(
               compilation,
               runtime_requirements,
               &dep.id,
@@ -517,7 +517,7 @@ impl DependencyTemplate for CommonJsExportRequireDependencyTemplate {
         UsedName::Inlined(inlined) => format!("{}{}", comment, inlined.render()),
       }
     } else {
-      module_raw(
+      compilation.runtime_template.module_raw(
         compilation,
         runtime_requirements,
         &dep.id,
