+++ crates/rspack_plugin_javascript/src/dependency/context/mod.rs
@@ -22,7 +22,7 @@ pub use require_resolve_context_dependency::{
 };
 use rspack_core::{
   ContextDependency, ContextMode, ContextOptions, DependencyRange, GroupOptions,
-  ResourceIdentifier, TemplateContext, TemplateReplaceSource, module_raw,
+  ResourceIdentifier, TemplateContext, TemplateReplaceSource,
 };
 
 fn create_resource_identifier_for_context_dependency(
@@ -92,7 +92,13 @@ fn context_dependency_template_as_require_call(
   } = code_generatable_context;
   let id = dep.id();
 
-  let mut expr = module_raw(compilation, runtime_requirements, id, dep.request(), false);
+  let mut expr = compilation.runtime_template.module_raw(
+    compilation,
+    runtime_requirements,
+    id,
+    dep.request(),
+    false,
+  );
 
   if compilation
     .get_module_graph()
@@ -124,7 +130,7 @@ fn context_dependency_template_as_id(
   } = code_generatable_context;
   let id = dep.id();
 
-  let expr = module_raw(
+  let expr = compilation.runtime_template.module_raw(
     compilation,
     runtime_requirements,
     id,
