+++ crates/rspack_plugin_javascript/src/dependency/esm/provide_dependency.rs
@@ -9,7 +9,7 @@ use rspack_core::{
   DependencyType, ExportsInfoGetter, ExtendedReferencedExport, FactorizeInfo, GetUsedNameParam,
   InitFragmentKey, InitFragmentStage, ModuleDependency, ModuleGraph, ModuleGraphCacheArtifact,
   NormalInitFragment, PrefetchExportsInfoMode, RuntimeSpec, SharedSourceMap, TemplateContext,
-  TemplateReplaceSource, UsedName, create_exports_object_referenced, module_raw,
+  TemplateReplaceSource, UsedName, create_exports_object_referenced,
 };
 use rspack_util::ext::DynHash;
 use swc_core::atoms::Atom;
@@ -199,7 +199,7 @@ impl DependencyTemplate for ProvideDependencyTemplate {
       format!(
         "/* provided dependency */ var {} = {}{};\n",
         dep.identifier,
-        module_raw(
+        compilation.runtime_template.module_raw(
           compilation,
           runtime_requirements,
           dep.id(),
