+++ crates/rspack_plugin_mf/src/sharing/provide_shared_module.rs
@@ -7,9 +7,8 @@ use rspack_core::{
   AsyncDependenciesBlock, AsyncDependenciesBlockIdentifier, BoxDependency, BuildContext, BuildInfo,
   BuildMeta, BuildResult, CodeGenerationResult, Compilation, ConcatenationScope, Context,
   DependenciesBlock, DependencyId, FactoryMeta, LibIdentOptions, Module, ModuleGraph,
-  ModuleIdentifier, ModuleType, RuntimeGlobals, RuntimeSpec, SourceType, async_module_factory,
-  impl_module_meta_info, impl_source_map_config, module_update_hash, rspack_sources::BoxSource,
-  sync_module_factory,
+  ModuleIdentifier, ModuleType, RuntimeGlobals, RuntimeSpec, SourceType, impl_module_meta_info,
+  impl_source_map_config, module_update_hash, rspack_sources::BoxSource,
 };
 use rspack_error::{Result, impl_empty_diagnosable_trait};
 use rspack_hash::{RspackHash, RspackHashDigest};
@@ -178,14 +177,14 @@ impl Module for ProvideSharedModule {
       .runtime_requirements
       .insert(RuntimeGlobals::INITIALIZE_SHARING);
     let factory = if self.eager {
-      sync_module_factory(
+      compilation.runtime_template.sync_module_factory(
         &self.get_dependencies()[0],
         &self.request,
         compilation,
         &mut code_generation_result.runtime_requirements,
       )
     } else {
-      async_module_factory(
+      compilation.runtime_template.async_module_factory(
         &self.get_blocks()[0],
         &self.request,
         compilation,
